Create readme
This project has let me explore several new technologies while creating my own personal website.  This is based on NextJS (in Typescript) using TailwindCSS for styling.  I have a long history of work with JS and React, but this is my first time with this stack. 

Will add more later